Move functions to schema module
